input: synaptics_dsx_i2c: implement double-tap to wake gesture
This feature allows to use the touchscreen as wakeup source.

When enabled, the touchscreen is never suspended and keeps listening
for finger touches. When a double-tap gesture is detected, a
KEY_POWER input event is sent to userspace.

The reporting mode is changed on suspend to reduce the power
consumption and restored on resume.

Change-Id: I697505f5a5d8f74b540c10fb49fbb970088c4390